setup rails API:
