--- app/controllers/application_controller.rb
@@ -1,4 +1,0 @@
-class ApplicationController < ActionController::Base
-  before_action :authenticate_user!
-  protect_from_forgery with: :exception
-end
