--- app/helpers/application_helper.rb
@@ -1,2 +1,0 @@
-module ApplicationHelper
-end
