+++ config/routes.rb
@@ -1,6 +1,10 @@
-Rails.application.routes.draw do
-  devise_for :users
-  # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
-  root  'posts#index'
-  resources :posts
+module 'namespace app'  
+  module 'namespace v1 '
+    Rails.application.routes.draw do
+      devise_for :users
+      # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
+      root  'posts#index'
+      resources :posts
+    end
+  end
 end
